Move logging to new project JsTestAdapter
