+++ JsTestAdapter/Logging/FileTestLogger.cs
@@ -7,7 +7,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace KarmaTestAdapter.Logging
+namespace JsTestAdapter.Logging
 {
     public class FileTestLogger : TestLoggerBase
     {
