+++ JsTestAdapter/Logging/TestLogLevel.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Web;
 
-namespace KarmaTestAdapter.Logging
+namespace JsTestAdapter.Logging
 {
     public enum TestLogLevel
     {
