+++ KarmaTestAdapter.sln
@@ -51,6 +51,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NpmProxy", "NpmProxy\NpmPro
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NpmProxyTests", "NpmProxyTests\NpmProxyTests.csproj", "{6109A014-AD9B-4EC0-AEBD-816FAECDA267}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "JsTestAdapter", "JsTestAdapter\JsTestAdapter.csproj", "{988CC656-0AF1-4F93-9D47-508FFE2877FA}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -81,6 +83,10 @@ Global
 		{6109A014-AD9B-4EC0-AEBD-816FAECDA267}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{6109A014-AD9B-4EC0-AEBD-816FAECDA267}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{6109A014-AD9B-4EC0-AEBD-816FAECDA267}.Release|Any CPU.Build.0 = Release|Any CPU
+		{988CC656-0AF1-4F93-9D47-508FFE2877FA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{988CC656-0AF1-4F93-9D47-508FFE2877FA}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{988CC656-0AF1-4F93-9D47-508FFE2877FA}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{988CC656-0AF1-4F93-9D47-508FFE2877FA}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
