+++ KarmaTestAdapter/Logging/KarmaServerTestLogger.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using JsTestAdapter.Logging;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
