+++ KarmaTestAdapterTests/KarmaTestAdapterTests.csproj
@@ -82,6 +82,10 @@
       <Project>{f15c2463-e70b-435a-9de6-69d6dedad66f}</Project>
       <Name>JsonServerClient</Name>
     </ProjectReference>
+    <ProjectReference Include="..\JsTestAdapter\JsTestAdapter.csproj">
+      <Project>{988cc656-0af1-4f93-9d47-508ffe2877fa}</Project>
+      <Name>JsTestAdapter</Name>
+    </ProjectReference>
     <ProjectReference Include="..\KarmaTestAdapter\KarmaTestAdapter.csproj">
       <Project>{7810c732-f9a2-40ec-910d-de3b9f50f8e1}</Project>
       <Name>KarmaTestAdapter</Name>
